requirements.txt
yapf>=0.31.0
pylint>=2.8.2
python-dotenv>=0.17.1
psycopg2-binary>=2.8.6
tqdm>=4.61.2
